requirements.txt
apex==0.9.10.dev0
h5py==3.8.0
joblib==1.2.0
matplotlib==3.7.1
numpy==1.23.4
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.0.1
pyrender==0.1.45
PyYAML==6.0
SciencePlots==2.1.0
scipy==1.9.1
smplx==0.1.28
timm==0.6.13
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.0
trimesh==3.21.5
